mise en place d'un système de gestion de l'application en fonction du mode d'environnement dans leque on se trouve (development ou production)

+++ src/components/DeleteArticle.js
@@ -1,20 +1,31 @@
 import axios from "axios";
 import React from "react";
 
-const DeleteArticle = ({id}) => {
-
-    const handleDelete = ( ) => {
-        axios.delete( "http://localhost:3003/articles/" + id );
-        window.location.reload();
+const DeleteArticle = ({ id, isDev, newsData, setNewsData }) => {
+  const handleDelete = () => {
+    if (isDev) {
+      axios.delete("http://localhost:3003/articles/" + id);
+      window.location.reload();
+    } else {
+      let Articles = newsData;
+      Articles = Articles.filter((article) => article.id !== id);
+      setNewsData(Articles);
     }
+  };
 
-    return (
-
-        <button onClick={ () => {
-            if ( window.confirm( "Voulez vous supprimer cette article" ) ) {
-                handleDelete();
-            }
-      }}>Delete</button>
+  return (
+    <button
+      onClick={() => {
+        if (isDev) {
+          if (window.confirm("Voulez vous supprimer cette article")) {
+            handleDelete();
+          }
+        } else {
+          handleDelete();
+        }
+      }}>
+      Delete
+    </button>
   );
 };
